requirements.txt
aruco-markers==1.1.6
attrs==26.1.0
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
numpy-quaternion==2024.0.13
opencv-contrib-python==4.13.0.92
packaging==26.0
pillow==12.2.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
reportlab==4.4.10
scipy==1.17.1
six==1.17.0
